@@ -1,36 +1,49 @@
 {
   "name": "outsera-avaliacao-frontend",
   "version": "1.0.0",
+  "private": true,
+  "engines": {
+    "node": "20.x"
+  },
   "scripts": {
     "start": "ng serve",
     "build": "ng build",
-    "test": "jest"
+    "test": "ng test"
   },
-  "private": true,
   "dependencies": {
-    "@angular/animations": "~17.0.0",
-    "@angular/common": "~17.0.0",
-    "@angular/compiler": "~17.0.0",
-    "@angular/core": "~17.0.0",
-    "@angular/forms": "~17.0.0",
-    "@angular/platform-browser": "~17.0.0",
-    "@angular/platform-browser-dynamic": "~17.0.0",
-    "@angular/router": "~17.0.0",
+    "@angular/animations": "~20.0.0",
+    "@angular/common": "~20.0.0",
+    "@angular/compiler": "~20.0.0",
+    "@angular/core": "~20.0.0",
+    "@angular/forms": "~20.0.0",
+    "@angular/platform-browser": "~20.0.0",
+    "@angular/platform-browser-dynamic": "~20.0.0",
+    "@angular/router": "~20.0.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~17.0.0",
-    "@angular/cli": "~17.0.0",
-    "@angular/compiler-cli": "~17.0.0",
-    "@types/jest": "^30.0.0",
-    "@types/mocha": "^10.0.10",
-    "@types/node": "^24.1.0",
-    "jest": "^29.7.0",
-    "jest-preset-angular": "^12.2.6",
-    "ts-jest": "^29.1.0",
-    "@angular-builders/jest": "^17.0.0",
-    "typescript": "~5.2.0"
+    "@angular-devkit/build-angular": "~20.0.0",
+    "@angular/animations": "~20.0.0",
+    "@angular/cli": "~20.0.0",
+    "@angular/common": "~20.0.0",
+    "@angular/compiler": "~20.0.0",
+    "@angular/compiler-cli": "~20.0.0",
+    "@angular/core": "~20.0.0",
+    "@angular/forms": "~20.0.0",
+    "@angular/platform-browser": "~20.0.0",
+    "@angular/platform-browser-dynamic": "~20.0.0",
+    "@angular/router": "~20.0.0",
+    "@types/jasmine": "^5.1.8",
+    "jasmine-core": "^5.9.0",
+    "karma": "^6.4.4",
+    "karma-chrome-launcher": "^3.2.0",
+    "karma-coverage": "^2.2.1",
+    "karma-jasmine": "^5.1.0",
+    "karma-jasmine-html-reporter": "^2.1.0",
+    "rxjs": "~7.8.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.15.0"
   }
 }
