@@ -1,7 +1,6 @@
-/// <reference types="jest" />
 import { TestBed } from '@angular/core/testing';
 import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
-import { MovieService } from '../../src/app/movie.service';
+import { MovieService, YearWins, StudioWins, ProducerInterval, Movie } from '../movie.service';
 
 describe('MovieService', () => {
   let service: MovieService;
@@ -25,35 +24,46 @@ describe('MovieService', () => {
     expect(req.request.params.get('size')).toBe('5');
     expect(req.request.params.get('year')).toBe('2020');
     expect(req.request.params.get('winner')).toBe('true');
+    req.flush({ content: [], totalPages: 1 });
   });
 
   it('deve listar anos com múltiplos vencedores', () => {
-    const mock = { years: [] };
-    service.getYearsWithMultipleWinners().subscribe();
+    const mock: YearWins[] = [];
+    service.getYearsWithMultipleWinners().subscribe(res => {
+      expect(res).toEqual(mock);
+    });
 
     const req = http.expectOne(r => r.url.includes('yearsWithMultipleWinners'));
-    req.flush(mock);
+    req.flush({ years: mock });
   });
 
   it('deve listar os principais estúdios', () => {
-    const mock = { studios: [] };
-    service.getTopStudios().subscribe();
+    const mock: StudioWins[] = [];
+    service.getTopStudios().subscribe(res => {
+      expect(res).toEqual(mock);
+    });
 
     const req = http.expectOne(r => r.url.includes('studiosWithWinCount'));
-    req.flush(mock);
+    req.flush({ studios: mock });
   });
 
   it('deve retornar intervalos de produtores', () => {
     const mock = { min: [], max: [] };
-    service.getProducerIntervals().subscribe();
+    service.getProducerIntervals().subscribe(res => {
+      expect(res).toEqual(mock);
+    });
 
     const req = http.expectOne(r => r.url.includes('maxMinWinIntervalForProducers'));
     req.flush(mock);
   });
 
   it('deve buscar vencedores por ano', () => {
-    service.getWinnersByYear(1984).subscribe();
+    const mock: Movie[] = [];
+    service.getWinnersByYear(1984).subscribe(res => {
+      expect(res).toEqual(mock);
+    });
     const req = http.expectOne(r => r.url.includes('winnersByYear'));
     expect(req.request.urlWithParams).toContain('year=1984');
+    req.flush(mock);
   });
 });
